requirements.txt
crewai 
unstructured 
pyowm 
tools 
wikipedia 
yfinance 
sec-api 
tiktoken 
faiss-cpu 
python-dotenv 
langchain-community 
langchain-core 
MarkdownTools

